Добавил экран с настройками. Сделал изменение темы для каждого контроллера через Core Data. Добавил статические методы для кодирования и раскодирования цвета. Рефакторинг

+++ Translator/Models/Settings/Settings+CoreDataProperties.h
@@ -7,8 +7,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSFetchRequest<Settings *> *)fetchRequest;
 
-@property (nonatomic) int16_t themeColor;
-@property (nonatomic) int16_t fontColor;
+@property (nonatomic) int32_t themeColor;
+@property (nonatomic) int32_t fontColor;
 @property (nonatomic) BOOL isRotate;
 @property (nonatomic) BOOL isTranslationByEnter;
